Updated sb-internet
sb-internet will now perform a port check to see if the port used by qbittorrent is open and portforwarded properly. It will show a check mark if the port is open and a red cross if it hasn't. This check is only performed when the qbittorrent client is running as a process.